requirements-dev.txt
# Development dependencies
-r requirements.txt

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
pytest-textual-snapshot>=0.4.0

# Linting and formatting
ruff>=0.1.0
mypy>=1.0.0

# Type stubs
types-setuptools

# Build tools
build>=0.10.0
twine>=4.0.0

# Documentation
sphinx>=6.0.0
sphinx-rtd-theme>=1.3.0

# Pre-commit hooks
pre-commit>=3.0.0
pre-commit-hooks>=4.0.0

# Security scanning
bandit>=1.7.0
safety>=2.3.0

# Dependency management
pip-tools>=7.0.0
pipdeptree>=2.0.0